refactoring ui element
