+++ workers/unity/Packages/com.improbable.gdk.core/Editor/CommonUIElements.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f88bced8bc86d43b2a873d14b93c10be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
