+++ workers/unity/Packages/com.improbable.gdk.mobile/Editor/Improbable.Gdk.Mobile.Editor.asmdef
@@ -2,6 +2,7 @@
     "name": "Improbable.Gdk.Mobile.Editor",
     "references": [
         "Improbable.Gdk.Core",
+        "Improbable.Gdk.Core.Editor",
         "Improbable.Gdk.Tools",
         "Improbable.Gdk.Mobile"
     ],
